Delete project's `ItemConsumer` with `Item`
Always delete the `ItemConsumer` of the project that owns the `Item`.
We do this before the `Item` deletion happens, so our soft-deletion
logic will hard delete if only that project had added the item.

Part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/575127

+++ ee/app/services/ai/catalog/items/base_destroy_service.rb
@@ -13,6 +13,9 @@ module Ai
           return error_no_permissions unless allowed?
           return error_no_item unless valid?
 
+          result = delete_item_consumer
+          return error(result.errors) if result.error?
+
           if delete_item
             track_ai_item_events('delete_ai_catalog_item', { label: item.item_type })
             return success
@@ -39,6 +42,14 @@ module Ai
 
         def error_response
           error(item.errors.full_messages)
+        end
+
+        def delete_item_consumer
+          consumer = project.configured_ai_catalog_items.for_item(item).first
+
+          return ServiceResponse.success unless consumer
+
+          Ai::Catalog::ItemConsumers::DestroyService.new(consumer, current_user).execute
         end
 
         def delete_item
